Add callbacks to Futures returned by Async
Success/Failure callbacks can now be added directly to the Future. eg

    Async.submit(() -> ...maybe returns a string...)
        .onSuccess(System.out::println)
        .onFailure(e -> logger.error("Couldn't do the thing", e));

The function passed to onSuccess is passed the result of the async task.
The function passed to onFailure is passed the exception thrown by the
async task (unless it is an ExecutionException where the underlying
cause will be passed).

Multiple callbacks can be added to the same future although there is no
guarantee as to the order of their execution.

A callback added to a scheduled task will be run when that task is
cancelled or throws an exception.

Change-Id: Ic6f5152f388f8e15a503284f7532980918edee39