DAQ-1258 Add JythonShellParser to enable multi line editing
When entering multiline commands in a remote GDA session, a new line is
automatically added rather than submitting the command as it is. This lets
previous lines be edited without aborting the input and starting again.
Once a command is complete, it is passed to the server.

If return is pressed while the cursor is not at the end of the command
(eg when editing a previoud command) a new line is entered rather than
submitting the full command. This is useful when editing but makes submitting
an edited command awkward. To prevent the need to move the cursor to the
end of the input to submit an edited command, Alt-Return will first move to
the end of the input and submit the command if possible.

Scrolling through history now scrolls through block at a time rather
than line at a time.

Using the up/down arrows in a remote session still moves by line but
using Ctrl-Up/Down will scroll by submitted command even if they have
multiple lines.

Change-Id: I27e546ae721b6525ef0d51cf4809a426598dacee