Add syntax highlighting to JythonShell
By default (or always if pygments is not available) there is no syntax
highlighting.

To enable highlighting, an environment variable needs to be set before
making the connection. The full command for this is

    # with default port and 'emacs' as the chosen theme
    GDA_THEME=emacs ssh -o 'SendEnv GDA_THEME' hostname -p2222

This can be simplified using the SendEnv argument in ~/.ssh/config eg:

    Host gda
        HostName hostname
        SendEnv GDA_THEME
        port 2222

Then connect using

    ssh gda

and GDA_THEME set in any of the standard places (.bashrc, .profile etc).

It is possible to change the default used when no variable is set by
setting the GDA property `gda.remote.colours` to one of the available
styles. If a default has been set, it is possible to disable colours
using the method above but passing 'none' as the theme.

Change-Id: Ie2ce9342285855d59b66232df3c1198c241f1bea