Suggest single tab when completing on empty line
Previously a tab was added to the existing line to create the suggestion.
For the client completion handling this worked but the shell handling
meant that multiple tabs were inserted when indenting on an empty line.

Completion is now a single tab and the 'before' field of the completion
suggestions holds the line so far (for the client to use as before).

Change-Id: Iafe1cccbc79a5eb47210ccc4b1a35ea3b2331896