B18-215 : Don't show old sample parameter settings in GUI if 'generic motors' parameters are available.
This change is to make the sample parameter GUI show settings consistent
with what will actually get used when running the scan. If generic motor
parameters are present in the sample parameter xml file, then only these are shown
in the GUI, and the old sample parameter settings (xy theta stage, user stage etc)
are hidden.

Change-Id: I33e3796dcac3f08cbeed6925d8e8b08935c1b02b