DAQ-1356 Remove InterpreterThread from ScriptProjectCreator
Do the check for existing interpreter in one of the class methods instead.
The additional thread doesn't seem to be required for correct behaviour.

Also make exiting interpreter search more comprehensive. Now it checks
for a Jython interpreter of the desired version instead of only checking
for an interpreter with a specific name. This is so that if a valid
interpreter is set up by some means other than by this class we don't
attempt to make another one.

In the same way change the nature assigned to projects so that it can
cope with an interpreter of a different name as long as it is the correct
jython version.

Change the logging level to info instead of error for the interpreter
check as it will be the logged every time a new client workspace is
launched when the check syntax option is enabled.

Change-Id: I84f928cf173f64f7f66ee1ba084ad4774b974793