Improve progress monitoring in StatusQueueView and warn about queue delays
Update the messages displayed as well as progress bar in the progress
monitor in the StatusQueueView, so it is easier to see what is happening.

Add code to warn about relevant processes which take too long to run.

Also add trace logging for various related method calls, with abridged
call stack for some methods which can be called via multiple routes on
both server and client.

Change-Id: Ic13f114cb09b4b9bb8fa35b8f06e0318032663de