Allow only a single StatusQueueView update job to run at once
With a large enough queue, an update can take so long to run that a new
job would be created before the last one finished, potentially leading
to a cascade failure.

For faster response with a large queue, store data in the content
provider in date order, rather than reverse date order (as it needs to
be displayed) and only reverse the order in getElements.

This allows updates to unknown beans to add the new bean to the end of
the list immediately, so they are displayed at the top of the view as
soon as possible, and can be then be updated as normal, while the view
is waiting for the full updated status to be retrieved from activemq.

Change-Id: Ic9b8b1f6b48389212ca5e92fc4e22711cd55aaa6