DAQ-644 Add sorted implementation of ScanDataPointCache
For some noisy motors, the positions can occasionally be out of order
causing processing to fail.
This adds an implementation of the ScanDataPointCache that maintains the
order of scan data points even if they are produced out of order.

The common behaviour between this and the existing implementation has been
extracted into an abstract super class

Change-Id: I3eb02ea16f6278dbfe1f16edf84030b7a69f7802