DAQ-1365 Modify the management of script project natures
This adds some reliability to the natures which are added to script
projects. The pydev specific nature should always be set to jython so
remove any previous Eclipse pydev nature first to allow it to be reset.

This will mean that if .py files are opened in the project later on
that the nature assigned to the project should always be jython.

Change-Id: I5a28bfae7c84e1515c1a98686d051998c9085560