[DAQ-1363] fix aborting scan prevents next scan from running
Removes the check for isActive() == false. The active flag is not actually
set to false until the next call to checkPaused, which does not happen
until the current scan is finished (note we are pausing the consumer
not the scan). Therefore any call to resume method that takes place
before this will be ignored. The awaitPaused flag will remain set
and cause the checkPaused method to pause the thread.

This has been replaced by checks to the awaitPaused flag, done
inside the stateChangeLock (could they be outside?).

Additionally rename FastRunCreator to DryRunProcessCreator to match
the name of the class of the createProcess method produces

Change-Id: I211339860f0fa86158b86c18ad46896545028e64