DAQ-1321 Remove functionality to import XML folder into workspace
This change removes: 
-XML project handling from ScriptProjectCreator
-Calls to these methods (from PreferenceInitializer and
ApplicationWorkbenchAdvisor) 
-The preference "Show all the XML in the GDA Configuration" from 
the GDA Scripts preferences page

The bl-config/xml folder is not used in the same way now as it 
might have been in the past. It is not clear why its contents would 
need to be opened within the client. Newer configs only contain
beamlinestaff and user_permissions files.

Change-Id: I13e931c9e43b4a32c96d337600658ade4788677e