Only insert a newline if editing a multiline command
When editing a command from a remote session, return is used for both
inserting a newline character and for submitting the command. This changes
the behaviour from only submitting the command if the cursor is at the
end of the command to also submitting the command if there are not already
multiple lines and the command is complete.

Change-Id: I5872cd1134c0d563fd78b0cdcc4c16769c6b6240