DAQ-1377 New path OverlapGridModel and Generator for ptychography
This path results in what is essentially the existing RasterModel, but
specified in terms of beam overlap. Beam dimensions are in the model to
keep the point generators free of scannable dependencies.

Overlap and beamsize are actually in AbstractOverlapModel; other paths
which are suitable for ptychography e.g. OverlapSpiral, OverlapRandom-
OffsetGrid can be added as children of this.

Change-Id: Ib11b3b93b7744ad679b27b52b969420dd1413ea6