DAQ-1370 Remove the CHECK_SCRIPT_SYNTAX preference
This preference used to have a large influence over the importing of
script projects and interpreter configuration in the client. This
behaviour has been redesigned in an attempt to make it more reliable.
This preference doesn't make sense and implications of setting it are
not clear by its name.

Remove calls obtaining this preference from the store and the use of it
in a property listening event.

In ScriptProjectCreator the code now follows the path that it would have
done if this was set to true.

Change-Id: Ida733c81bf0480ac8a25f1eafd113d1496a6467c