DAQ-1329 Add new scannable BeamDimensions
This is a very simple scannable to represent beam size in 2D. It will be
needed later on to define scan paths in terms of beam overlap (important
for ptychography).

No GUI is provided as it is not intended for users; beamline scientists
can set the dimensions through the Jython console, or provide reasonable
values for Spring initialisation.

Change-Id: I5f4f8a55616ea49d2a34c27e3305830d6743277e