DAQ-1377 Create path editor for OverlapGridModel
This new path editor specifies the model's beam overlap through either
a textbox or slider. It also includes the standard snake and continuous
controls.

In this change, the IEventService injection is moved up the hierarchy
from AbstractRegionEditor to AbstractModelEditor so that it can be used
in AbstractPathEditor to achieve the following:

On creation of this editor, we look for a BeamDimensions scannable with
name "beamDimensions"; the dimensions of this are passed to the path
model.

Change-Id: I5a2428c6daf88e146ccfc3acfffa4b062ca91dd4