Fix ScannableGroupTest.createAndCheckScannableGroupContaining()
This function would previously accept a variable number of Scannables
and create a ScannableGroup from them. However, subsequent checks
would fail unless the ScannableGroup contained exactly the
scannable s1 created in the setup() function. Fortunately, this
is the only way it is ever called.

This change modifies the function so it accepts only one Scannable
and checks the contents of the ScannableGroup against the Scannable
actually passed.


Change-Id: Ib3b19a17460c859dbab8e5b27e939e057270b0b7