[DAQ-1156] Add DummyXspress3Detector
This was previously defined as a Jython object in i24's localStation.py

This Java version can be created in Spring, thus allowing us to delete
MapFactory, as the only remaining use in the codebase was to make
the Jython DummyXspress3Detector findable.

This version differs in two ways:
- It does not contain a (Dummy)Xspress3Controller, as this was not
called by the Jython object.
- It does not hard code a name: this can be set in Spring or allowed
to default to the bean id

Change-Id: I3965e1430ba9af376e21ff2cf298963dcfc6acc6