gda_servers_core: make 'server started' check work on RHEL 7
On RHEL 6 "nc" is the old BSD netcat tool. It accepts a zero timeout to the
"-w" option.

On RHEL 7, "nc" is the nmap ncat tool. It doesn't allow a zero timeout.

To cope with the different implementations, instead use the "-z" (zero-I/O)
option (and no timeout) to test the GDA server. We don't actually need a
timeout any more, because we test using the new status port (not telnet/ssh),
and the server accepts and immediately closes connections to its status port.

Change-Id: Icf637a3d64effa5a263aa3772d6e61f298e202bc