Add Alt-Enter prompt to JythonShell
When a multiline command is being edited, enter inserts a new line
instead of submitting the command. There is also the option to use
Alt-enter to directly submit the command without having to move to the
end of the buffer first. This key binding is not immediately obvious so
this adds a reminder to the right of the screen when a new line would be
inserted.

Change-Id: I08f39d7364c60614ee6e2464a262880aec2ad9b1