DAQ-1427 Add null check to JythonShell
When a remote connection is closed without sending Ctrl-d (closing the window
or xkill etc), jline's readline returns null. This was previously converted to
a string "null" which was run on the server.

If this did not cause any terminal output, JythonShell would enter a loop of
repeatedly running the "null" command

Change-Id: I4a4310f07c9f13df6cc02ca0a24dbd3b4dfcd1e8