I20-338 : FluorescenceDetector view : check number of detector elements when 'applying' settings
This commit adds a check to FluorescenceDetector view to ensure that when
'applying' detector settings, the number of detector elements in the settings
is consistent with the number of elements on the detector. A warning dialog
is displayed if there is a mismatch. Previously, the error message and content
in log when this problem occurred made it hard for user to diagnose and
fix the problem.

Change-Id: I8a50c272a8d89d9fe6527d02e295a63218087b50