move filtering out of Malcolm controlled scannables to AcquisitionDevice
This partially reverses the change with git id 69ea27e, except that
before that change, we got the scannables for each name first, then
filtered them if they were malcolm controlled. This caused a
ScanningException if there was no GDA scannable for a particular name.
With this change we filter the names first.

It's better to get the names in AcquistionDevice rather than ScanProcess
as AcquisitionDevice still would do that if the names weren't set, and
this would fail if any of the axes were malcolm-controlled. This was
the case in tests that use Malcolm devices, e.g. MalcolmStaticScanTest
which uses a DummyMalcolmDevice. The previous change broke those tests,
this change fixes them.

Change-Id: Ic1b674c40ac50e55535042d917ef93d543d9c598