Make NudgePositionerComposite more robust when no motor limits are set
If no limits are set in Epics for a motor, the functions
getLowerInnerLimit() & getUpperInnerLimit() in ScannableMotor
will return null.

This is returned (via Jython) to
NudgePositionerComposite.determineScannableLimits() as 'None'.
The function tries to parse this to a double, throws an exception
which is uncaught and the Live Controls view is not displayed properly.

This change catches and logs the exception and returns zero for the
limits. GDA may still not be able to move the motor because of the
zero limit, but the rest of the controls in the view will be displayed
and usable.

Change-Id: Ie8bfee2463c67f911ba99af54c5a5b28401deafc