Delete NameMappedEpicsPneumaticCallback.initializationCompleted()
It is almost identical to the base class function it overrides.

The base class function additionally clears the existing
position arrays, which is probably unnecessary, as the object
will presumably just have been created, but (for the same reason)
will do no harm.

Also delete the logger, which is no longer used as a result
of this change.

Change-Id: I9833dfd25cd85c67701e4307f241a31fc5adecdf