Stop LivePlotViewManager implementing IScanDataPointObserver
LivePlotViewManager is used to ensure that a plot view is open at the
start of a scan. It is currently added to the JSF as a ScanDataPoint observer
and then actively ignores ScanDataPoints. It only receives the updates it
needs because the JSF passes all updates to SDP observers.

This changes it to be added as a JSF observer. This should prevent problems if
JSF is ever improved to only pass requested updates. There is also a (very)
minor performance improvement as data points don't have to be handled.

Change-Id: Ie8458aa37da03251e026f3b2c814f82b09518c12