[DAQ-1448] Add tooltips to queue view
This is the first part of making it easier to see queue entries.
It allows the contents of each cell to be displayed as a tooltip.

Currently, the tooltip text is wrapped at 40 characters. There is
a bug in Eclipse Mars which means that the tool tip flickers
and disappears if it extends beyond the edge of the screen.

We believe that this is fixed in later versions of the Eclipse
platform. For the time being, wrap the tool tip to make this
less likely to happen.

Change-Id: I3eafc702dde53a97568f6316a2af4cb65a16b5a7