[DAQ-1453] close MessageConsumer when ConsumerImpl is paused
add add tests for remove/replace/reorder with and without a
running ConsumerImpl

This fixes a bug that broke removing, reordering or pausing a bean
that is on the submission queue (i.e. hasn't started running yet).

We need to close the MessageConsumer before a new MessageConsumer is
created/used by the AbstractQueueConnection.remove/replace/reorder
method due to the prefetch feature of ActiveMQ making this beans not
available to be consumed by another consumer (despite the fact that
a QueueBrowser can see them, e.g. see the implemenation of 
AbstractQueueConnection.remove())

Change-Id: Icb1860809601d71344a37d9ec90184a944be80de