Move pause logic into scan when its enclosing script is paused
When a script pause is requested but the flag is not checked by the
script, a scan started from the script should wait for the script to be
resumed before starting.

With the previous implementation, checking for script pauses at the start of
the scan set the script status to paused on the server. This could lead
to an unresumable scan if the the script pause is requested using the pause
button on the client.

1. Start script
	(script: running, scan: idle)
2. Pause button requests script pause (but script doesn't check flag)
	(script: running, scan: idle)
3. Scan starts from script, checks script pause flag and pauses script
	(and waits)
	(script: paused, scan: running)
4. Pause button pressed again (to resume). Scan is running so gets
	paused. Scan still waiting for script to resume.
	(script: paused, scan: paused)
5. Pause button pressed again. Scan is paused so gets resumed but still
	waits for script to be resumed.
	(script: paused, scan: running)
6. Scan pause/resume ad infinitum

This changes the scan to 'steal' the paused status from the script
rather than waiting for the script to be resumed. If the scan is not
started from a script, the script pause flag is ignored as before.

Change-Id: I48859ba7ca33cade036552b6cbd72de09acd5db3