[DAQ-1406] fix being able to reorder the queue more than once
The problem was that if reorder() was called on an ISubmitter, after
reading in the beans, reordering them, it used that submitter itself
to submit them, inside a try-with-resources block, meaning that the
SubmitterImpl's disconnect method is called. This closed the submitter's
MessageProducer but didn't nullify it. The next time reorder was called
an IllegalStateException when the bean was submitted.

This fix makes the following changes 
1. SubmitterImpl.disconnect now nullifies the MessageProducer after it
has been closed so that a closed one won't be used.
2. AbstractQueueConnection.doReorder and doReplace now both always
use a new submitter to re-submit the reordered beans. Even if this is
an ISubmitter, we don't know for sure that it's a submitter for the
same queue, since the reorder and replace methods in
AbstractQueueConnection are public and implement methods declared in
IQueueConnection, as well as being called by SubmitterImpl.reorder
and replace which call these methods with a queue name of their
submission queue name.

This problem shows that the queue management methods, such as 
reorder, remove, replace, cleanQueue and clearQueue, and even getQueue,
don't belong on a class that is actually a Submitter or a Consumer.
Instead they should be on their own interface, e.g. IQueueManager.
This should be dealt with as part of DAQ-1464


Change-Id: Ic138989c04f79348646243fa9c628ba80c62694a