[DAQ-1457] fix RemoteScannableServiceTest
the bug is that events are not received by the proxy scannable when
subtractFive is called after addFive. This is order dependent, it will
always fail on all but the first test to require beans to be sent
and received correctly. This is because
IEventService.createRemoteService actually caches the remote service
rather than always creating a new one. However, it does create a new one
if the service has been disconnected.

This bug was introduced by renaming the method from isDisconnected to
isConnected without changing the return value. The set method was still
called setDisconnected. 

The fix is to now return the correct value in isConnected and rename
setDisconnected to setConnected, reversing the argument in any
invocation of this method.

Change-Id: Iba67d1af69c0e80d296aceed5b771bb847d14a8d