B18-54 : NexusTreeWriter for buffered Xspress2 detector
A new NexusTreeWriter class has been added -- this can be used to add data
to a Nexus file using LazyWritableDataset based methods and provides an
alternative to using NexusDataWriter to add the data :
- Add data to be written by calling 'addData' with  an array of NexusTreeProvider
objects; datasets are extracted from this and stored internally.
- Remove data from NexusTreeProvider by calling removeNodeFromNexusTree
(this ensures that NexusDataWriter will not also write the same data).
- Write data to Nexus file by calling 'writeNexusData'. This will write
all the data stored in NexusTreeWriter by appending data to LazyWritableDatasets
and then clear it. On first call to writeNexusData, the Nexus file will
be created (if necessary), and LazyWritableDatasets will be initialised.
Note :
 -'addData' can be called multiple times before writing the data.
 - The writer can output data to a different file from the gda scan Nexus file;
 links will be added to datasets in external Nexus in the scan Nexus file if needed.
 - Compression level for the datasets can be set using setCompressionLevel
 (default = 1, LZW compression).

Change-Id: I8316a9e7ab3d1cc959f9e16b6da86d6619332dbd