Only enable "Show Details..." button if pressing it will do something
If more than one job is selected, only the first is considered, and
even then it is only shown if it is in the run list (i.e. the job is
either running or has completed running) so only enable when only a
single item is selected in the run list.

This prevents Unhandled event loop exceptions when the button is pressed
with no selection or when the first selection is in the submitted list.

Change-Id: I1421385ad2020577748b0fdb241eaff1ea004021