DAQ-1484 Make AbstractConnection fields private
This is to protect the internal state of the fields and allow them to be
synchronized. Hopefully this will fix errors seen on Jenkins like:

Cannot subscribe to topic org.eclipse.scanning.request.acquire.topic
with URI vm://localhost?broker.persistent=false

This can happen when disconnect is called but the corresponding session,
connection or queueSession is not nulled.

And also NPE which can occur when a create method is called
simultaneously with the disconnect method.

Change-Id: I83bb25e05b5726d3aafdffb40f77fbe20140ddd7