I18-212 Add associated axis move on double-click in Mapping
If the mapping stage manager has an associated axis, double-clicking on 
the MappedData view will trigger a search in the top-most nexus file for
a per-scan monitor of this axis name. If found, it will be included in
the stage move i.e. the move consists of fast axis to x coordinate of
double click, slow axis to y, and associated axis to its position in the
nexus file.

 - The UI thread is now only used to display the confirmation message.


Change-Id: I600bd78460b999d680df78f106b3410419e4641b