[DAQ-1410] fire MalcolmEvents from MalcolmDevice instead of positionComplete
MalcolmDevice now fires a MalcolmEvent to its IMalcolmEventListeners
instead of calling positionComplete on its positionCompleteListeners

MalcolmDevice never really needed to iterate through points of the
inner scan as it overwrote the stepIndex of the IPosition with that
received by malcolm, and this was the only field read.

Also fix DummyMalcolmDevice to do the same.

Change-Id: I822e34ebda56d22ee273f416c93b298ad9d16e15