DAQ-1451 Add Basic Roi/Scanpath/Mutator handler classes
These enums represent the Rois, Scanpaths (covering an area) and
Mutators to be specified in an MScan. In the case of Roi and
AreaScanpath, their instances link to the corresponding IROI and *Model
classes used in new scanning, allowing them to be constructed
conveniently once the required enum instance has been specified. These
instance 'types' will be supplied as part of the array of objects that
describes the scan from Jython, ready constructed so that there is no
need for string parsing on the Java side to establish what type of
model/roi is required.

Change-Id: Ib8b49efb77f2723bda7187ac438823016a96b519