[DAQ-1485] scanning changes for new malcolm version
The new malcolm version has the attribute 'simultaneousAxes' instead
of 'axesToMove'. GDA needs to be able to handle both attributes
(for the moment). We can treat these attributes the same.

For the new malcolm version, we need to set axesToMove in the model
that we configure malcolm with all the time. Previously it was optional.
This change handles this by setting this value all the time.

Change-Id: Icc34f98efd2659319fd2c74e79c37d94eee236f0