RmiAutomatedExporter - Resolve class loading problem
In some cases objects could not be exported over RMI as the classes
could not be seen by the Spring RmiServiceExporter. Here this is
resolved by setting the thread context classloader to the uk.ac.gda.core
OSGi bundle loader before starting the exports. It is then reset after
the exports are done.

Add exception to logging during shutdown.

Change-Id: If4c715041be1b0a8ce8a08ee2890ff61bd9c3593