DAQ-1482 Allow Spring Profiles to be specified at gda server startup
Richard has already added the capability for the SpringObjectServer to
load configs using Spring Profiles via a gda.spring.profiles.active
property. This change allows the property to be set via the gda script,
which will override any settings in the property file, using the
--springprofiles option. Once we have got rid of legacy use of GDA
profiles, this could become the function of the existing -p option.

Change-Id: I8074259dad04af763d86b0f996443902ff8650ed