[DAQ-1520] don't close the session in updateQueue
as we now use the session returned by getSession which is in use
by the thread running the main loop. This cause the next call to
consumeNextMessage() to fail, a catch block calls disconnect().
This sets lots of things to null that we still need if we don't exit
the main read loop, which we don't

Change-Id: I565702469fac4f7715b67adb2922c04f930a304a