RmiAutomatedExporter - Use Spring bundle classloader not core
This was a mistake in a previous change. As core is a buddy of Spring
anything that core can see Spring also can however classes in bundles
which are direct buddies of Spring would not be visible with the core
classloader but should be with Spring. Additionally the Spring bundle
classloader is the one used when the beans are instantiated so if that
has worked this should too.

Change-Id: I745ca2d32d68a41c806dcbe91f6e08189d77b1b1