B18-239 : Fix for 'pause after repetition' in command queue for Xas scans
This fixes a problem for Xas scans (e.g. qexafs) where 'pause after repetition'
could not be exited. This problem was due to the script status being left in a
'paused' state if ScriptBase.checkForPauses() is exited early due to an
InterruptedException being thrown. This occurs if 'abort current task
and pause queue' or 'stop current task and skip to next' buttons are
pressed in the command queue view when paused at the end of a scan.


Change-Id: I41feec1b554620e7196627cab036375ba28bafa7