DAQ-1178 Removes OSGi static bind methods
With the OSGi runtime provided in the Eclipse Oxygen.3 TP static OSGi
service bind methods are not allowed.

The solution employed here is to remove the static modifier. This will
now work but it is messy. It would be nicer to convert the code to use
activators and remove this "ServiceHolder" pattern.

Change-Id: I36ccaee1355ad85135f961f782abf9f64092c58d