[DAQ-1554] remove unneccessary HealthAttribute class
this class is just the same as StringAttribute. We now create a
StringAttribute whether the meta id from malcolm is that for
String or Health (see NTScalarDeserializer). In malcolm v2 the meta id
will be that for Health, but for v3 it will be that for String. Once
we no longer wish to support malcolm v2, we can remove the contant
StringAttribute.HEALTH_ID, see DAQ-1555.

Change-Id: Iafeed8a4e573d07bc67ca5be4f9790f5504ed083