[DAQ-1436] remove unnecessary MalcolmUtil class
This class was used to get device state and health from a MalcolmMessage
It was written as if it wasn't clear how the device state or health
would be contained within the MalcolmMessage, specifically within the
value attribute. In fact where we call this method we will always have
a ChoiceAttribute for getState, or a StringAttribute for getHealth.
This reduced the complexity of getting these values to such a degree
that the code to do this can be inlined and this class removed.

Change-Id: I11e7129524b2815cd5b9fcc8a5de41e9a6e58e36