DAQ-1178 This fixes tests after OSGi static bind methods are removed
The solution used is to create a new ServiceTestHelper which provides a
way to setup all the required services and injects them into instances
of the "ServiceHolders" this works in the same way OSGi does in the real
application.

This provides for significant simplification of the tests which
previously included lots of repeated code to setup services.

This also remove an xcen tests from scanning tests. This is needed
because core test depends on scanning test. Without this change core
test would also need to depend on scanning.example.xcen

Change-Id: I4de535d4036dd878fc5cf0253990fd57830817db