DAQ-1451 Scan Clause Element Processors and Context object
These Classes handle the processing of the various elements of a Scan
Clause. A Scan Clause is the bit of an MScan command that describes the
Scannables being moved, the Roi that encapsulates the motion and the
AreaScanpath that is to be followed (including any Mutators to the path).
As such an actual scan could include several Scan Clauses depending on
its dimensionality. A Scan clause does not include any Detectors or
Monitors that are enabled during the scan as these are handled
separately in the new scanning infrastructure.

The ClauseContext is a form object that gets filled in by the processors,
one for each element of the clause. This allows the forthcoming Scan
Builder object to just loop over the set of processors calling the
process method for each of the elements, with the type specific handling
being encapsulated within the appropriate methods on the Clause Context
object. These also handle the validation of whether elements can be set
based on the state of what has already been filled in. The ClauseContext
object also requires a validate step to be called before the set values
can be retrieved from it. Additionally, this step makes any adjustments
that are necessary to translate between the clause syntax and the way
the DAWN Model classes represent things.

The ClauseContext also defines the overall clause grammar which the
processors use to determine whether their element type is valid at the
current place in the clause.

At this stage the Mutator support is only a placeholder with minimal
validation as the underlying links to the models need to change so that
we don't need a new one every time we apply a Mutator to it. This will
follow later.

For an area based clause (Mapping scan) the acceptable syntax is:
1. Two scannables followed by
2. A Roi followed by
3. Its required number of params followed by
4. An AreaScanpath followed by
5. Its required number of params followed by
6. An optional Mutator

The Roi and AreaScanpath can be defaulted (to Rectangle and Raster) by
omitting them.
Some Rois (e.g. Polygon) do not have an upper bound for their number of
points (except that of the variable type) and so these require the
AreaScanpath to be specified or else the clause cannot be successfully
interpreted.

Other classes, including the one that constructs the set of processors
for each clause will follow this, but this set should be testable as a
discrete chunk of functionality simplifying the reviewing process.

Change-Id: Ida59d1d826f3a697cf6670ed82e475a94b23ce94