[DAQ-1517] remove IAttributableDevice and use specific methods instead
IMalcolmDevice should encapsulate the way that it uses attributes to
communicate with the EpicsV4ConnectionService. Client code does not
need access to arbitrary attributes.

Attributes were previously used on the client to get the axesToMove
and dataset attributes of malcolm devices. This used attributes in the
DeviceRequest sent to and received back from the server. This change
replaces attributes in DeviceRequest and DeviceInformation with
specific fields for axes to move and datasets.

Note, this is bit clunkly but is unavoidable due to the lack of
scalability of the proxy mechanism. Hopefully this part of the change
is only temporary until the underscore proxies, e.g. _RunnableDevice
are replaced with JMS proxies or some other mechanism.

Change-Id: I295656cb452b8b81f582ea3ce84474c3d02c32d8