DAQ-704 Add scannable overwriting bypass
Overwriting scannables should be prevented where possible but it is
useful/required in places leading to various workaround being used.

This adds a context manager to explicitly state that overwriting a
scannable is required and blocks attempts to overwrite them otherwise.

To ease the transition from the previous workarounds, this currently
only logs a warning but should be updated to throw an Exception once
the new context manager is being used.

Use:
    >>> a = DummyScannable('a')
    >>> a = 3
    # throws NameError "Can't overwrite scannable"
    >>> with overwriting:
    ...     a = 3
    ...
    >>> a
    3
    >>>

'overwriting' is added as a builtin, meaning that if it is overwritten
itself, it can be restored either by running a full `reset_namespace` or
by using

    >>> overwriting = __builtins__.overwriting

Change-Id: I214ba935b0422a16b426a7676760409cbe29b973