[DAQ-980] Improve error messaging in LevelRunner
- After running the devices in a level by calling eservice.invokeAll(),
check first for abortException i.e. for an exception explicitly
thrown by a device. This should be thrown in preference to assuming
that a cancelled future means a timeout has occurred.

- Remove await() from the finally block in run(). If there is
a failure in scanning (especially a device timeout), the call
to awaitQuiescence() in await() may itself throw an exception,
which will hide the original exception. I can't see any reason
to wait for the service to stop if we are aborting the scan anyway,
so this change means that await() is called only if
the scan has succeeded.

- if there is an exception due to timeout, list in the exception
messages the devices that have timed out

Also:

- removed toString(devices), as it was used only to construct an
error message, and it did more of less the same thing
as the default toString() for List<>.

- this also required removing the overridden version of
toString(devices) in ScannablePositioner

The tests have been refactored so that there is now only one level
runner. The delay in scanning is now specified in the individual
scannable, and exception throwing by a parameter on the runner.

Change-Id: Ibbfbcf88d044616892717683f4b0f61bd0acf9bd