DAQ-1586 Finder - Deprecate searching for objects by interface string
There are now methods to find objects by class object instead of String.
These methods are type safe and don't require lots of string
manipulation.

Initially this deprecation is to prevent new code using these methods
but ideally they can be removed in the future. Logging has been added to
help identify when they are used. Once these are removed the Finder can
be significantly simplified as much of the code is to support this
String manipulation.

Change-Id: If5fa895802ee5025176698c2a85f682d7456728d