DAQ-1574 Fix error messages in StatusPanel
updateStatusLabel() can either show the message supplied, or (more com-
monly), can update a) the number of points, b) the total exposure time,
and c) the smallest step size in the current path.

Every element in the Mapping perspective that can change the path calls
this method after receiving an update. At startup, this throws a lot of
errors since there is no path.

This change adds an early return in the latter scenario when the result
of the path calculation is null.

Change-Id: I22ba4006257fa2838935d2b6af2c79bb60ac86b5