[DAQ-1317] remove exitProcess field from KillBean
This doesn't just stop the consumer as the message shown to the user
in ConsumerView implies, it actually kills the server VM. This should
not be possible, and simply sending the bean to the server without
this field has the desired effect.

The exit() method is left in ConsumerImpl() and is still called if
a process doesn't respond after one day. This may be removed in a
future change.

Change-Id: I045a8727038cae9e796a871c500235d25ced9811