Stop icon tooltips disappearing when a scan is running
Forcing updates on the ApplicationActionBarAdvisor has the side effect of
removing any currently displayed tooltip (only with Oxygen target platform).
This makes it very hard/impossible to see tooltips in gda when a scan is
running. Setting force to false fixes the issue, with scan status information
still apparently updated correctly as the scan progresses.

Change-Id: I7869e5f557efd35343642db4e2abc870479c49a1